+++ lib/flutter_network.dart
@@ -1,3 +1,5 @@
+/// Package build upon powerful HTTP client [dio](https://pub.dev/packages/dio),
+/// that extends some basic functionality and simplifies client-server logic handling process.
 library flutter_network;
 
 export 'src/connection/connection.dart';