+++ lib/src/connection/connection.dart
@@ -1,6 +1,7 @@
 import 'package:connectivity_plus/connectivity_plus.dart';
 
-/// Providing internet connection state for the [NetworkService]
+/// Providing internet connection state for the [NetworkService].
+/// Wrapper above [Connectivity] from connectivity_plus package.
 ///
 /// Use [isConnected] and [isNotConnected] static fields to quickly get
 /// the status of the Internet connection
@@ -9,12 +10,6 @@ import 'package:connectivity_plus/connectivity_plus.dart';
 class Connection {
   const Connection({Connectivity? connectivity}) : _connectivity = connectivity;
 
-  factory Connection.instance() {
-    return _instance ??= const Connection();
-  }
-
-  static Connection? _instance;
-
   final Connectivity? _connectivity;
 
   /// You can use this [Connectivity] instance for your own purpose
@@ -22,17 +17,17 @@ class Connection {
 
   /// Asynchronously provides internet connection state
   ///
-  /// [true] - has connection
+  /// If true - has connection
   ///
-  /// [false] - no connection
+  /// If false - no connection
   Future<bool> get isConnected async =>
       await connectivity.checkConnectivity() != ConnectivityResult.none;
 
   /// Asynchronously provides internet connection state
   ///
-  /// [true] - no connection
-  /// [false] - has connection
+  /// If true - no connection
   ///
+  /// If false - has connection
   Future<bool> get isNotConnected async =>
       await connectivity.checkConnectivity() == ConnectivityResult.none;
 }