+++ lib/src/data/data.dart
@@ -1,4 +1,4 @@
-export 'network_error/_network_error.dart';
+export 'network_error/network_error.dart';
 export 'network_error/connection_network_error.dart';
 export 'network_error/dio_network_error.dart';
 export 'network_error/response_network_error.dart';