+++ lib/src/data/network_error/connection_network_error.dart
@@ -1,5 +1,6 @@
-import '_network_error.dart';
+import 'network_error.dart';
 
+/// Represents [NetworkError] returned, when there is no internet connection.
 class ConnectionNetworkError extends NetworkError {
   const ConnectionNetworkError();
 }