+++ lib/src/data/network_error/dio_network_error.dart
@@ -1,12 +1,15 @@
-import 'package:dio/dio.dart' show DioError;
+import 'package:dio/dio.dart' show Response, DioException;
 
-import '_network_error.dart';
+import 'network_error.dart';
 
+/// Describes [NetworkError] when request ended with the error.
+/// Usually returned when [Response.statusCode] is >= 300 < 400 or >= 500.
 class DioNetworkError extends NetworkError {
-  final DioError dioError;
+  /// Refer to [DioException] from dio package.
+  final DioException dioException;
 
   const DioNetworkError({
     required super.stackTrace,
-    required this.dioError,
+    required this.dioException,
   });
 }