+++ lib/src/data/network_error/response_network_error.dart
@@ -1,8 +1,11 @@
 import 'package:dio/dio.dart' show Response;
 
-import '_network_error.dart';
+import 'network_error.dart';
 
+/// Describes [NetworkError] when request ended with the error.
+/// Usually returned when [Response.statusCode] is >= 400 < 500.
 class ResponseNetworkError extends NetworkError {
+  /// Refer to [Response] from dio package.
   final Response response;
 
   const ResponseNetworkError({