+++ lib/src/data/network_error/type_network_error.dart
@@ -1,5 +1,6 @@
-import '_network_error.dart';
+import 'network_error.dart';
 
+/// Describes [NetworkError] when response data wasn't decoded because of [TypeError].
 class TypeNetworkError extends NetworkError {
   final TypeError typeError;