+++ lib/src/data/result/error_data.dart
@@ -1,3 +1,6 @@
+/// Represents error response data for [NetworkService] when decoding error [Response.data].
+///
+/// Create your own error data subclasses by extending or implementing this class.
 abstract class ErrorData {
   const ErrorData();
 }