+++ lib/src/data/result/error_result.dart
@@ -1,10 +1,13 @@
-import 'package:flutter_network/src/data/network_error/_network_error.dart';
+import 'package:dio/dio.dart' show Response;
 
-import 'result.dart';
-import 'error_data.dart';
+import '../data.dart';
 
+/// Describes the error response done by [NetworkService].
+/// For successful response refer to [SuccessfulResult].
 class ErrorResult<T, E extends ErrorData> implements Result<T, E> {
+  ///
   final NetworkError error;
+  /// Decoded error [Response.data].
   final E? data;
 
   const ErrorResult({required this.error, this.data});