+++ lib/src/data/result/result.dart
@@ -1,14 +1,19 @@
+import 'package:dio/dio.dart' show Response;
+
 import '../data.dart';
 
-/// The base class returned by the network service [RectClient].
+/// The base abstract class returned by the [NetworkService].
+/// For detailed info about subclasses, refer to [SuccessfulResult] and [ErrorResult].
 ///
 /// [T] is a generic type for decoding [Response] data.
 ///
-/// [SuccessfulResult] - successful response with decoded data of [T] type.
-///
-/// [ErrorResult] — error response with specified [Exception].
+/// [E] is a generic type for decoding error [Response] data and must be
+/// subclass of the [ErrorData].
 abstract class Result<T, E extends ErrorData> {
-  const factory Result.success({required final T? data}) = SuccessfulResult;
+  const factory Result.success({
+    required final Response response,
+    required final T? data,
+  }) = SuccessfulResult;
 
   const factory Result.error({
     required final NetworkError error,