+++ test/logger_test.dart
@@ -1,6 +1,7 @@
-import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:dio/dio.dart';
 import 'package:http_mock_adapter/http_mock_adapter.dart';
+import 'package:connectivity_plus/connectivity_plus.dart';
 
 import 'package:flutter_network/flutter_network.dart';